Switch to Faraday
Switch to using Faraday to make requests to MinFraud. This makes use of
Net::HTTP::Persistent to share connections in a connection pool.
The goal is to increase the robustness of the client and as a bonus,
speed up response times.

Added Minfraud::MinfraudClient as a way to send requests.
The send_transaction method takes in a Minfraud::Transaction and
returns a Minfraud::Response.

Removed configuration requirements for the Minfraud module.

Made license_key a required field on Minfraud::Transaction.

Updated version to 1.0.0.